Reorganize stuff for better testability.
